refactor: 事件绑定弹窗提取面板左右两部分组件
