+++ packages/settings/events/meta.js
@@ -1,5 +1,6 @@
 import { commonEvents } from './src/commonjs/events.js'
-import BindEventsMethods from './src/components/BindEventsMethods.vue'
+import BindEventsDialogBodyLeft from './src/components/BindEventsDialogBodyLeft.vue'
+import BindEventsDialogBodyRight from './src/components/BindEventsDialogBodyRight.vue'
 
 export default {
   id: 'engine.setting.event',
@@ -12,7 +13,7 @@ export default {
     commonEvents
   },
   components: {
-    // 事件绑定的函数列表组件
-    BindEventsMethods
+    BindEventsDialogBodyLeft,
+    BindEventsDialogBodyRight
   }
 }
